backend/requirements.txt

fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
python-pptx==0.6.21
python-docx==0.8.11
openai==1.3.0
